+++ src/io/github/alexeygrishin/btree/TreeNameHelper.java
@@ -2,7 +2,7 @@ package io.github.alexeygrishin.btree;
 
 public interface TreeNameHelper {
 
-    public String getFullName(int dataId);
+    public String getFullName(long dataId);
 
     public String truncate(String fullName, int targetLen);
